integration updates

+++ task-manager-app/src/main/java/com/fsd/taskManager/repository/ParentTaskRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.fsd.taskManager.entity.ParentTask;
 import com.fsd.taskManager.entity.Task;
 
-public interface ParentTaskRepository extends JpaRepository<ParentTask, Integer>{
+public interface ParentTaskRepository extends JpaRepository<ParentTask, Long>{
 
 	public ParentTask findByParentTask(String parentTask);
 }
