+++ task-manager-app/src/main/java/com/fsd/taskManager/repository/TaskRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.fsd.taskManager.entity.Task;
 
-public interface TaskRepository extends JpaRepository<Task, Integer> {
+public interface TaskRepository extends JpaRepository<Task, Long> {
 
 	@Query(value = "Select t From Task t where t.task=?1 or t.parentTask=?2 or t.priority=?3")
 	public List<Task> search(String task, String parentTask, int priority);
